+++ app/controllers/admin/invites_controller.rb
@@ -1,0 +1,4 @@
+module Admin
+  class InvitesController < Admin::ReadonlyController
+  end
+end