+++ app/controllers/categories_controller.rb
@@ -1,6 +1,6 @@
 class CategoriesController < ApplicationController
   def show
     @category = Category.friendly.find(params[:slug])
-    bustable_fresh_when(@category)
+    with_caching(@category)
   end
 end