+++ app/forms/asset_form.rb
@@ -1,0 +1,4 @@
+class AssetForm < Reform::Form
+  property :asset_file
+  validates :asset_file, presence: true
+end