+++ app/models/template.rb
@@ -1,6 +1,0 @@
-class Template < ApplicationRecord
-
-  attachment :preview_image
-
-  validates :name, uniqueness: { case_sensitive: false }
-end