+++ config/routes.rb
@@ -34,6 +34,7 @@ Rails.application.routes.draw do
   resource :feedback, controller: 'feedback'
 
   namespace :editorial do
+    resources :assets, only: [:index, :new, :create]
     resources :news, only: [:index, :new, :edit, :update]
     get '/:section/news/:slug' => 'news#show', as: :news_article
 
@@ -61,11 +62,13 @@ Rails.application.routes.draw do
     # keep these alphabetically sorted
     # code order determines order in the UI
     resources :agencies
+    resources :assets
     resources :categories
     resources :custom_template_nodes
     resources :departments
     resources :feedbacks
     resources :general_contents
+    resources :invites
     resources :ministers
     resources :news_articles
     resources :nodes
@@ -96,6 +99,11 @@ Rails.application.routes.draw do
 
   resources :departments, only: :index
   resources :ministers, only: :index
+  resources :invites, only: [:show, :create, :new, :edit, :update] do
+    collection do
+      get :required
+    end
+  end
 
   get 'categories/:slug' => 'categories#show', as: :category