+++ spec/fabricators/invite_fabricator.rb
@@ -1,0 +1,3 @@
+Fabricator(:invite) do
+  email { Fabricate.sequence(:email) { |i| "user-#{i}@some-agency.gov.au" } }
+end