+++ CHANGELOG.md
@@ -7,9 +7,12 @@ This file is influenced by http://keepachangelog.com/.
 - Added ability to link to another page by id
 - Added EDITOR.md as a central location for documenting the editor
 - News lists can be filtered by ministers and departments
+- Image upload function available in editorial
+- Toggle to require all visitors to have an invite
 
 ### Changed
 - Updated ui-kit to v1.8.0
+- Now adds Cache-Control headers (public, max-age 5 mins) to supplement ETAGs.
 
 ### Fixed
 - Pluralise 'ministers' only where appropriate